Refactoring snapshot taking again

+++ Lombiq.Tests.UI/Services/SqlServerManager.cs
@@ -115,7 +115,8 @@ public sealed class SqlServerManager : IAsyncDisposable
         string snapshotDirectoryPathRemote,
         string snapshotDirectoryPathLocal = null,
         string containerName = null,
-        bool useCompressionIfAvailable = false)
+        bool useCompressionIfAvailable = false,
+        int maxRetries = 3)
     {
         var filePathRemote = GetSnapshotFilePath(snapshotDirectoryPathRemote);
         var filePathLocal = GetSnapshotFilePath(snapshotDirectoryPathLocal ?? snapshotDirectoryPathRemote);
@@ -134,29 +135,47 @@ public sealed class SqlServerManager : IAsyncDisposable
             useCompressionIfAvailable &&
             (server.EngineEdition == Edition.EnterpriseOrDeveloper || server.EngineEdition == Edition.Standard);
 
-        var backup = new Backup
-        {
-            Action = BackupActionType.Database,
-            CopyOnly = true,
-            Checksum = true,
-            Incremental = false,
-            ContinueAfterError = false,
-            // We don't need compression for setup snapshots as those backups will be only short-lived and we want them
-            // to be fast.
-            CompressionOption = useCompression ? BackupCompressionOptions.On : BackupCompressionOptions.Off,
-            SkipTapeHeader = true,
-            UnloadTapeAfter = false,
-            NoRewind = true,
-            FormatMedia = true,
-            Initialize = true,
-            Database = _databaseName,
-        };
+        var retryCount = 0;
 
-        var destination = new BackupDeviceItem(filePathRemote, DeviceType.File);
-        backup.Devices.Add(destination);
-        // We could use SqlBackupAsync() too but that's not Task-based async, we'd need to subscribe to an event which
-        // is messy.
-        backup.SqlBackup(server);
+        do
+        {
+            // We need to recreate the backup object and set its devices each time we retry, because the SMO API
+            // disposes them after the backup completes.
+            var backup = new Backup
+            {
+                Action = BackupActionType.Database,
+                CopyOnly = true,
+                Checksum = true,
+                Incremental = false,
+                ContinueAfterError = false,
+                // We don't need compression for setup snapshots as those backups will be only short-lived and we want
+                // them to be fast.
+                CompressionOption = useCompression ? BackupCompressionOptions.On : BackupCompressionOptions.Off,
+                SkipTapeHeader = true,
+                UnloadTapeAfter = false,
+                NoRewind = true,
+                FormatMedia = true,
+                Initialize = true,
+                Database = _databaseName,
+            };
+
+            var destination = new BackupDeviceItem(filePathRemote, DeviceType.File);
+            backup.Devices.Add(destination);
+            // We could use SqlBackupAsync() too but that's not Task-based async, we'd need to subscribe to an event
+            // which is messy.
+            backup.SqlBackup(server);
+
+            await Task.Delay(1000);
+            retryCount++;
+        }
+        while (!File.Exists(filePathLocal) && retryCount < maxRetries + 1);
+
+        if (!File.Exists(filePathRemote))
+        {
+            throw new FileNotFoundException(
+                $"Failed to create snapshot file at \"{filePathRemote}\" after {maxRetries.ToTechnicalString()} retries.");
+        }
+
         if (!string.IsNullOrEmpty(containerName))
         {
             if (File.Exists(filePathLocal)) File.Delete(filePathLocal);
